fix(ios): register preferences branch files in Xcode project and fix build error
Add 6 missing Swift files from preferences branch to project.pbxproj
(UserPreferences, PreferencesService, SettingsStore, SettingsViewModel,
SettingsView, UserAvatarView). Fix ambiguous operator in
FullScreenImageViewer.swift.

+++ apps/ios/Listwell/Listwell/Views/Detail/FullScreenImageViewer.swift
@@ -13,7 +13,7 @@ struct FullScreenImageViewer: View {
     var body: some View {
         ZStack {
             Color.black
-                .opacity(1 - min(abs(dragOffset.height) / 400, 0.5))
+                .opacity(1 - min(abs(dragOffset.height) / CGFloat(400), 0.5))
                 .ignoresSafeArea()
 
             VStack(spacing: 0) {
